2.8 library - write_bin_build_wine - Do not remove links to $HOME for scripts targeting library version 2.7 or older

+++ play.it-2/lib/libplayit2.sh
@@ -33,7 +33,7 @@
 ###
 
 library_version=2.8.1~dev
-library_revision=20180505.9
+library_revision=20180506.1
 
 # set package distribution-specific architecture
 # USAGE: set_architecture $pkg
@@ -2521,13 +2521,18 @@ write_bin_build_wine() {
 	  mkdir --parents "${WINEPREFIX%/*}"
 	  # Use LANG=C to avoid localized directory names
 	  LANG=C wineboot --init 2>/dev/null
-	  # Remove most links pointing outside of the WINE prefix
-	  rm "$WINEPREFIX/dosdevices/z:"
-	  find "$WINEPREFIX/drive_c/users/$(whoami)" -type l | while read directory; do
-	    rm "$directory"
-	    mkdir "$directory"
-	  done
 	EOF
+
+	if ! ( [ $version_major_target -lt 2 ] || [ $version_minor_target -lt 8 ] ); then
+		cat >> "$file" <<- 'EOF'
+		  # Remove most links pointing outside of the WINE prefix
+		  rm "$WINEPREFIX/dosdevices/z:"
+		  find "$WINEPREFIX/drive_c/users/$(whoami)" -type l | while read directory; do
+		    rm "$directory"
+		    mkdir "$directory"
+		  done
+		EOF
+	fi
 
 	if [ "$APP_WINETRICKS" ]; then
 		cat >> "$file" <<- EOF
