+++ play.it-2/src/30_launchers_wine.sh
@@ -66,13 +66,18 @@ write_bin_build_wine() {
 	  mkdir --parents "${WINEPREFIX%/*}"
 	  # Use LANG=C to avoid localized directory names
 	  LANG=C wineboot --init 2>/dev/null
-	  # Remove most links pointing outside of the WINE prefix
-	  rm "$WINEPREFIX/dosdevices/z:"
-	  find "$WINEPREFIX/drive_c/users/$(whoami)" -type l | while read directory; do
-	    rm "$directory"
-	    mkdir "$directory"
-	  done
 	EOF
+
+	if ! ( [ $version_major_target -lt 2 ] || [ $version_minor_target -lt 8 ] ); then
+		cat >> "$file" <<- 'EOF'
+		  # Remove most links pointing outside of the WINE prefix
+		  rm "$WINEPREFIX/dosdevices/z:"
+		  find "$WINEPREFIX/drive_c/users/$(whoami)" -type l | while read directory; do
+		    rm "$directory"
+		    mkdir "$directory"
+		  done
+		EOF
+	fi
 
 	if [ "$APP_WINETRICKS" ]; then
 		cat >> "$file" <<- EOF
